backEnd/requirements.txt
asgiref==3.5.2
cffi==1.15.1
dj-database-url==1.0.0
Django==3.2.16
django-cors-headers==3.13.0
django-on-heroku==1.1.2
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.3.0
psycopg2-binary==2.9.3
pycparser==2.21
pytz==2022.2.1
sqlparse==0.4.2
typing-extensions==4.4.0
tzdata==2022.2
whitenoise==6.2.0
